PR3a: Split MomentumEquation into Equations/Momentum/{base,standard,mixed}.py
- Extract LinearMomentumBase (ABC, 884 lines) → Equations/Momentum/base.py
- Extract LinearMomentum (concrete, 367 lines) → Equations/Momentum/standard.py
- Extract LinearMomentumMixed (concrete, 379 lines) → Equations/Momentum/mixed.py
- Fix relative import paths (3-level depth: from ...Materials.Material, from ..base, etc.)
- Create Equations/Momentum/__init__.py with re-exports
- Convert old MomentumEquation.py to shim (one-line imports from new locations)
- Fix bug: ConvergenceCriteria.py TYPE_CHECKING import (was non-relative 'from MomentumEquation')
- Fix bug: Output/SaveFields.py TYPE_CHECKING import (wrong relative depth)
- Delete stray Equations/Equations/ nested dir (pre-existing placeholder)

Verification (pre-commit):
- Clean import of safeincave works
- Old paths (from safeincave.MomentumEquation import LinearMomentum) work
- New paths (from safeincave.Equations.Momentum.standard import LinearMomentum) work
- No import cycles

Next: Complete ConvergenceCriteria.py → Simulation/Convergence/ split + Simulators.py → Simulation/Simulators/ split
