+++ safeincave/Equations/Momentum/__init__.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Linear momentum equation solvers (displacement-based and mixed u-p)."""
+
+from .base import LinearMomentumBase  # noqa: F401
+from .standard import LinearMomentum  # noqa: F401
+from .mixed import LinearMomentumMixed  # noqa: F401
+
+__all__ = ["LinearMomentumBase", "LinearMomentum", "LinearMomentumMixed"]
